+++ scripts.d/99-finalize.sh
@@ -14,9 +14,10 @@ ffbuild_dockerbuild() {
     set -e
     mkdir -p "$FFBUILD_DESTPREFIX"
 
-    if [[ $TARGET == linux* ]]; then
-        rm "$FFBUILD_DESTPREFIX"/lib/lib*.so* || true
-        rm "$FFBUILD_DESTPREFIX"/lib/*.la || true
+    if [[ "${PREFER_SHARED}" != "1" ]]; then
+        if [[ $TARGET == linux* ]]; then
+            rm "$FFBUILD_DESTPREFIX"/lib/lib*.so* || true
+            rm "$FFBUILD_DESTPREFIX"/lib/*.la || true
+        fi
     fi
-
 }
